// src/types/shopify.d.ts
interface ShopifyImage {
  url: string;
  altText: string;
  width: number;
  height: number;
}

interface ShopifyPrice {
  amount: string;
  currencyCode: string;
}
// ...
export interface ShopifyOrderDetailed {
  name: string;
  orderNumber: number;
  processedAt: string;
  financialStatus: string;
  fulfillmentStatus: string;
  totalPriceV2: ShopifyPrice;
  subtotalPriceV2: ShopifyPrice;
  totalShippingPriceV2: ShopifyPrice;
  shippingAddress: {
    address1: string;
    address2: string | null;
    city: string;
    province: string;
    zip: string;
    country: string;
  };
  lineItems: {
    edges: {
      node: {
        title: string;
        quantity: number;
        variant: {
          title: string;
          image: ShopifyImage;
        };
        originalTotalPrice: ShopifyPrice;
      };
    }[];
  };
}